% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-workbook.R
\name{bdc_build_workbook}
\alias{bdc_build_workbook}
\title{Uses bookdown to build the workbook inside a specified folder}
\usage{
bdc_build_workbook(
  stage_folder = tempdir(),
  source = system.file("workbook", package = "bigdataclass"),
  db_folder = "database",
  file_folder = "files",
  book_folder = "books"
)
}
\arguments{
\item{stage_folder}{The folder where the book will be compiled}

\item{source}{The path of the workbook}

\item{db_folder}{Location of the source 'database' folder}

\item{file_folder}{Location of the source 'files' folder}

\item{book_folder}{Location of the source 'books' folder}
}
\description{
Uses bookdown to build the workbook inside a specified folder
}
