% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{bdc_create_dates}
\alias{bdc_create_dates}
\title{Returns a tibble with date values}
\usage{
bdc_create_dates(
  days_in_segment = 10,
  no_of_segments = 100,
  start_date = "2016-01-01"
)
}
\arguments{
\item{days_in_segment}{Number of days in each segment}

\item{no_of_segments}{Number of segments to be created}

\item{start_date}{Start date}
}
\description{
Returns a tibble with date values
}
