% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-project.R
\name{bdc_create_project}
\alias{bdc_create_project}
\title{Creates a new RStudio project with the exercises}
\usage{
bdc_create_project(
  folder = "big-data-class",
  src = system.file("exercises", package = "bigdataclass"),
  overwrite = FALSE,
  db_connection = "con <- connection_open(RSQLite::SQLite(), 'database/local.sqlite')",
  dbi_connection = "con <- dbConnect(RSQLite::SQLite(), 'database/local.sqlite')",
  files_path = "files",
  books_path = "books"
)
}
\arguments{
\item{folder}{Path to folder where to build the project}

\item{src}{Path to the source folder of the exercises}

\item{overwrite}{Overwrite class folder if it exists}

\item{db_connection}{Text code for connecting to the database}

\item{files_path}{Location of the files folder}

\item{books_path}{Location of the books folder}
}
\description{
Creates a new RStudio project with the exercises
}
