% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{bdc_create_view}
\alias{bdc_create_view}
\title{Creates a view in a database based on a tbl query}
\usage{
bdc_create_view(query, name)
}
\arguments{
\item{query}{A tbl based query}

\item{name}{The name to give the new database view}
}
\description{
Creates a view in a database based on a tbl query
}
