% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{bdc_create_view_lineitems}
\alias{bdc_create_view_lineitems}
\title{Creates the line items view}
\usage{
bdc_create_view_lineitems(con, name = "v_lineitems")
}
\arguments{
\item{con}{Database connection}

\item{name}{Defaults to "v_lineitems"}
}
\description{
Creates the line items view
}
