% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/views.R
\name{bdc_create_view_orders}
\alias{bdc_create_view_orders}
\title{Creates the orders view}
\usage{
bdc_create_view_orders(con, name = "v_orders")
}
\arguments{
\item{con}{Database connection}

\item{name}{Defaults to "v_orders"}
}
\description{
Creates the orders view
}
