% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{bdc_db_create_dates}
\alias{bdc_db_create_dates}
\title{Creates a table with multiple date values}
\usage{
bdc_db_create_dates(con, start_date = "2016-01-01")
}
\arguments{
\item{con}{Database connection}

\item{start_date}{Start date}
}
\description{
Creates a table with multiple date values
}
