% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookups.R
\name{bdc_db_lookups}
\alias{bdc_db_lookups}
\title{Creates the product and customer tables}
\usage{
bdc_db_lookups(
  con,
  product_data = bdc_create_products(),
  customer_data = bigdataclass::customers
)
}
\arguments{
\item{con}{Database connection}

\item{product_data}{Data for the products table}

\item{customer_data}{Data for the customer table}
}
\description{
Creates the product and customer tables
}
