% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders.R
\name{bdc_db_orders}
\alias{bdc_db_orders}
\title{Creates the orders and line items table in a given database connection}
\usage{
bdc_db_orders(
  con,
  avg_daily_orders = 100,
  avg_no_items = 3,
  days_in_segment = 10,
  no_of_segments = 100,
  seed = 7878
)
}
\arguments{
\item{con}{Database connection}

\item{avg_daily_orders}{Number of daily orders}

\item{avg_no_items}{Number of average number of items per order}

\item{days_in_segment}{Number of days inside a segment}

\item{no_of_segments}{Number of segments}

\item{seed}{Seed to randomize}
}
\description{
Creates the orders and line items table in a given database connection
}
