% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-books.R
\name{bdc_init_books}
\alias{bdc_init_books}
\title{Creates and populates the 'books' folder}
\usage{
bdc_init_books(folder = "books")
}
\arguments{
\item{folder}{Defaults to 'books'}
}
\description{
Creates and populates the 'books' folder
}
