% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-local.R
\name{bdc_init_class_local}
\alias{bdc_init_class_local}
\title{Creates a new RStudio project with the exercises and needed data sets}
\usage{
bdc_init_class_local(
  folder = "big-data-class",
  avg_daily_orders = 100,
  avg_no_items = 3,
  days_in_segment = 10,
  no_of_segments = 100,
  seed = 7878,
  product_data = bdc_create_products(),
  customer_data = bigdataclass::customers,
  start_date = "2016-01-01",
  no_transactions = 50000,
  batch_size = 5000,
  no_files = 5,
  src = system.file("exercises", package = "bigdataclass"),
  overwrite = FALSE,
  include_books = FALSE
)
}
\arguments{
\item{folder}{Path to folder where to build the project}

\item{avg_daily_orders}{Number of average daily orders}

\item{avg_no_items}{Average number of sales in an order}

\item{days_in_segment}{Days in a segment}

\item{no_of_segments}{Number of segments}

\item{seed}{Seed to randomize to}

\item{product_data}{Product data frame source}

\item{customer_data}{Customer data frame source}

\item{start_date}{The start date for the orders}

\item{no_transactions}{Number of transactions to save to file}

\item{batch_size}{File batch size}

\item{no_files}{Number of files}

\item{src}{Path to the source folder of the exercises}

\item{overwrite}{Overwrite class folder if it exists}

\item{include_books}{Download and save the book data}
}
\description{
Creates a new RStudio project with the exercises and needed data sets
}
