% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-database.R
\name{bdc_init_database}
\alias{bdc_init_database}
\title{Initializes the database with the orders data}
\usage{
bdc_init_database(
  con,
  avg_daily_orders = 100,
  avg_no_items = 3,
  days_in_segment = 10,
  no_of_segments = 100,
  seed = 7878,
  product_data = bdc_create_products(),
  customer_data = bigdataclass::customers,
  start_date = "2016-01-01",
  orders_view = "v_orders",
  lineitems_view = "v_lineitems"
)
}
\arguments{
\item{con}{Database connection}

\item{avg_daily_orders}{Number of average daily orders}

\item{avg_no_items}{Average number of sales in an order}

\item{days_in_segment}{Days in a segment}

\item{no_of_segments}{Number of segments}

\item{seed}{Seed to randomize to}

\item{product_data}{Product data frame source}

\item{customer_data}{Customer data frame source}

\item{start_date}{The start date for the orders}

\item{orders_view}{Name to give to the orders view}

\item{lineitems_view}{Name to give the line items view}
}
\description{
Initializes the database with the orders data
}
