% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-files.R
\name{bdc_init_files}
\alias{bdc_init_files}
\title{Initializes the files based on a database connection}
\usage{
bdc_init_files(
  con,
  folder = "files",
  no_transactions = 50000,
  batch_size = 5000,
  no_files = 5
)
}
\arguments{
\item{con}{Database connection}

\item{folder}{Folder path where to create the files}

\item{no_transactions}{Number of transactions to save to file}

\item{batch_size}{File batch size}

\item{no_files}{Number of files}
}
\description{
Initializes the files based on a database connection
}
