% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-db.R
\name{bdc_init_sqlite}
\alias{bdc_init_sqlite}
\title{Creates an empty SQLite database}
\usage{
bdc_init_sqlite(folder = "database")
}
\arguments{
\item{folder}{Folder location to place the database file}
}
\description{
Creates an empty SQLite database
}
