% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bdc_libraries}
\alias{bdc_libraries}
\title{Returns a vector with a list of libraries used in the workbook}
\usage{
bdc_libraries(files = book_files())
}
\arguments{
\item{files}{The full path of all of the Rmd files}
}
\description{
Returns a vector with a list of libraries used in the workbook
}
