% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bdc_outline}
\alias{bdc_outline}
\title{Returns a CAT outline of the class, based on the Rmd's titles}
\usage{
bdc_outline(files = book_files())
}
\arguments{
\item{files}{The full path of all of the Rmd files}
}
\description{
Returns a CAT outline of the class, based on the Rmd's titles
}
