% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{book_files}
\alias{book_files}
\title{Returns a list of all the Rmd files inside the workbook folder}
\usage{
book_files(book_path = system.file("workbook", package = "bigdataclass"))
}
\arguments{
\item{book_path}{Location of the workbook folder}
}
\description{
Returns a list of all the Rmd files inside the workbook folder
}
