% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-customers.R
\docType{data}
\name{customers}
\alias{customers}
\title{List of customers}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 90 rows and 6 columns.}
\usage{
customers
}
\description{
List of customers
}
\keyword{datasets}
