% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-db.R
\name{bdc_init_db}
\alias{bdc_init_db}
\title{Initializes schema, tables and views}
\usage{
bdc_init_db(con)
}
\arguments{
\item{con}{PostgreSQL database connection}
}
\description{
Initializes schema, tables and views
}
